Do not rely on route status when finalizing routes
It can happen that the cfRoute status does not have an FQDN at the time
it gets deleted. Therefore, we should always build the FQDN from the
current route and its domain.

Also, fix CFDomains finalization to not let the domain go while there
are related routes around. Therefore when finalizing routes we could safely
assume that the domain exists and we can calculate its FQDN name.

+++ controllers/controllers/networking/cfdomain_controller.go
@@ -88,6 +88,14 @@ func (r *CFDomainReconciler) finalizeCFDomain(ctx context.Context, log logr.Logg
 		return ctrl.Result{}, err
 	}
 
+	if len(domainRoutes) == 0 {
+		if controllerutil.RemoveFinalizer(cfDomain, CFDomainFinalizerName) {
+			log.V(1).Info("finalizer removed")
+		}
+
+		return ctrl.Result{}, nil
+	}
+
 	for i := range domainRoutes {
 		err = r.client.Delete(ctx, &domainRoutes[i])
 		if err != nil {
@@ -96,11 +104,7 @@ func (r *CFDomainReconciler) finalizeCFDomain(ctx context.Context, log logr.Logg
 		}
 	}
 
-	if controllerutil.RemoveFinalizer(cfDomain, CFDomainFinalizerName) {
-		log.V(1).Info("finalizer removed")
-	}
-
-	return ctrl.Result{}, nil
+	return ctrl.Result{Requeue: true}, nil
 }
 
 func (r *CFDomainReconciler) listRoutesForDomain(ctx context.Context, cfDomain *korifiv1alpha1.CFDomain) ([]korifiv1alpha1.CFRoute, error) {
